// lib/api/generated/model/testItemInput.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */
import type { TestItemInputCustomValues } from "./testItemInputCustomValues";

/**
 * テスト項目の内容。
 */
export interface TestItemInput {
  id: string;
  /** @minimum 0 */
  position?: number;
  /**
   * @minLength 1
   * @maxLength 100
   */
  code: string;
  /** @maxLength 20000 */
  target_feature?: string;
  is_spacer?: boolean;
  pattern_table_id?: string | null;
  /** @maxLength 20000 */
  viewpoint?: string;
  /** @maxLength 20000 */
  content?: string;
  /** @maxLength 20000 */
  preconditions?: string;
  /** @maxLength 20000 */
  test_data?: string;
  /** @maxLength 20000 */
  steps?: string;
  /** @maxLength 20000 */
  expected_result?: string;
  /** @maxLength 20000 */
  notes?: string;
  custom_values?: TestItemInputCustomValues;
}